move errors to their own files
