+++ tests/pytest_tests/system_tests/test_launch/test_launch.py
@@ -6,7 +6,7 @@ import wandb
 from wandb.errors import CommError
 from wandb.sdk.internal.internal_api import Api as InternalApi
 from wandb.sdk.launch.launch import run
-from wandb.sdk.launch.utils import LaunchError
+from wandb.sdk.launch.errors import LaunchError
 
 
 def test_launch_incorrect_backend(runner, user, monkeypatch, wandb_init, test_settings):
