+++ tests/pytest_tests/unit_tests_old/tests_launch/test_launch_aws.py
@@ -8,7 +8,7 @@ import wandb.sdk.launch._project_spec as _project_spec
 import wandb.sdk.launch.launch as launch
 from wandb.sdk.launch.environment.aws_environment import AwsEnvironment
 from wandb.sdk.launch.runner.sagemaker_runner import SagemakerSubmittedRun
-from wandb.sdk.launch.utils import LaunchError
+from wandb.sdk.launch.errors import LaunchError
 
 from tests.pytest_tests.unit_tests_old.utils import fixture_open
