+++ tests/pytest_tests/unit_tests_old/tests_launch/test_launch_gcp.py
@@ -7,7 +7,7 @@ import wandb.sdk.launch.launch as launch
 from google.cloud import aiplatform
 from wandb.sdk.launch.environment.gcp_environment import GcpEnvironment
 from wandb.sdk.launch.runner.vertex_runner import get_gcp_config, run_shell
-from wandb.sdk.launch.utils import LaunchError
+from wandb.sdk.launch.errors import LaunchError
 
 from .test_launch import mock_load_backend, mocked_fetchable_git_repo  # noqa: F401
