+++ tests/pytest_tests/unit_tests_old/tests_launch/test_launch_kubernetes.py
@@ -11,7 +11,8 @@ from wandb.sdk.launch.runner.kubernetes_runner import (
     MAX_KUBERNETES_RETRIES,
     maybe_create_imagepull_secret,
 )
-from wandb.sdk.launch.utils import LaunchError, make_name_dns_safe
+from wandb.sdk.launch.errors import LaunchError
+from wandb.sdk.launch.utils import make_name_dns_safe
 
 from .test_launch import mock_load_backend, mocked_fetchable_git_repo  # noqa: F401
