+++ wandb/cli/cli.py
@@ -36,10 +36,9 @@ from wandb.sdk.artifacts.artifacts_cache import get_artifacts_cache
 from wandb.sdk.launch.launch_add import _launch_add
 from wandb.sdk.launch.sweeps import utils as sweep_utils
 from wandb.sdk.launch.sweeps.scheduler import Scheduler
+from wandb.sdk.launch.errors import ExecutionError, LaunchError
 from wandb.sdk.launch.utils import (
     LAUNCH_DEFAULT_PROJECT,
-    ExecutionError,
-    LaunchError,
     check_logged_in,
     construct_launch_spec,
 )
