+++ wandb/sdk/launch/builder/build.py
@@ -28,11 +28,10 @@ from .._project_spec import (
     LaunchProject,
     fetch_and_validate_project,
 )
+from ..errors import ExecutionError, LaunchError
 from ..utils import (
     LAUNCH_CONFIG_FILE,
     LOG_PREFIX,
-    ExecutionError,
-    LaunchError,
     resolve_build_and_registry_config,
 )
 from .abstract import AbstractBuilder
