+++ wandb/sdk/launch/builder/noop.py
@@ -4,7 +4,7 @@ from typing import Any, Dict
 from wandb.sdk.launch.builder.abstract import AbstractBuilder
 from wandb.sdk.launch.environment.abstract import AbstractEnvironment
 from wandb.sdk.launch.registry.abstract import AbstractRegistry
-from wandb.sdk.launch.utils import LaunchError
+from wandb.sdk.launch.errors import LaunchError
 
 from .._project_spec import EntryPoint, LaunchProject
