+++ wandb/sdk/launch/environment/gcp_environment.py
@@ -3,7 +3,7 @@ import logging
 import os
 import re
 
-from wandb.sdk.launch.utils import LaunchError
+from wandb.sdk.launch.errors import LaunchError
 from wandb.util import get_module
 
 from .abstract import AbstractEnvironment
