+++ wandb/sdk/launch/environment/local_environment.py
@@ -1,7 +1,7 @@
 """Dummy local environment implementation. This is the default environment."""
 from typing import Any, Dict, Union
 
-from wandb.sdk.launch.utils import LaunchError
+from wandb.sdk.launch.errors import LaunchError
 
 from .abstract import AbstractEnvironment
