+++ wandb/sdk/launch/github_reference.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from typing import Optional, Tuple
 from urllib.parse import urlparse
 
-from wandb.sdk.launch.utils import LaunchError
+from wandb.sdk.launch.errors import LaunchError
 
 PREFIX_HTTPS = "https://"
 PREFIX_SSH = "git@"
