+++ wandb/sdk/launch/launch.py
@@ -12,12 +12,11 @@ from ._project_spec import create_project_from_spec, fetch_and_validate_project
 from .agent import LaunchAgent
 from .builder.build import construct_builder_args
 from .runner.abstract import AbstractRun
+from .errors import ExecutionError, LaunchError
 from .utils import (
     LAUNCH_CONFIG_FILE,
     LAUNCH_DEFAULT_PROJECT,
     PROJECT_SYNCHRONOUS,
-    ExecutionError,
-    LaunchError,
     construct_launch_spec,
     validate_launch_spec_source,
 )
