+++ wandb/sdk/launch/launch_add.py
@@ -6,10 +6,10 @@ import wandb.apis.public as public
 from wandb.apis.internal import Api
 from wandb.sdk.launch._project_spec import create_project_from_spec
 from wandb.sdk.launch.builder.build import build_image_from_project
+from wandb.sdk.launch.errors import LaunchError
 from wandb.sdk.launch.utils import (
     LAUNCH_DEFAULT_PROJECT,
     LOG_PREFIX,
-    LaunchError,
     construct_launch_spec,
     validate_launch_spec_source,
 )
