+++ wandb/sdk/launch/registry/google_artifact_registry.py
@@ -4,7 +4,7 @@ import re
 from typing import Tuple
 
 from wandb.sdk.launch.environment.gcp_environment import GcpEnvironment
-from wandb.sdk.launch.utils import LaunchError
+from wandb.sdk.launch.errors import LaunchError
 from wandb.util import get_module
 
 from .abstract import AbstractRegistry
