+++ wandb/sdk/launch/registry/local_registry.py
@@ -1,8 +1,8 @@
 """Local registry implementation."""
 import logging
 from typing import Tuple
-
-from wandb.sdk.launch.utils import LaunchError, docker_image_exists
+from wandb.sdk.launch.errors import LaunchError
+from wandb.sdk.launch.utils import docker_image_exists
 
 from ..environment.abstract import AbstractEnvironment
 from .abstract import AbstractRegistry
